cpm_etl_script-master/requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
dnspython==2.2.1
et-xmlfile==1.1.0
greenlet==2.0.0.post0
idna==3.4
Naked==0.1.32
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
pycparser==2.21
pycryptodome==3.15.0
pymongo==4.3.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.43
urllib3==1.26.12
